Base Parametrization uphill & change vertDist to float

+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/CostCalculatorShortestDist.java
@@ -7,7 +7,7 @@ public class CostCalculatorShortestDist implements IfCostCalcHeuristic{
     }
 
     @Override
-    public double calcCosts(double dist, double verDist) {
+    public double calcCosts(double dist, float verDist) {
         return dist + 0.0001;
     }
 }
