+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/IfCostCalculator.java
@@ -1,6 +1,6 @@
 package mg.mgmap.activity.mgmap.features.routing.profile;
 
 public interface IfCostCalculator {
-    public double calcCosts(double dist, double verDist);
+    public double calcCosts(double dist, float verDist);
 
 }
